Move the lading page to route /home and do formating and responsive for the different screen size
