+++ client/src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Navbar from "../components/navbar/navbar.svelte";
-  import Footer from "../components/footer/footer.svelte";
+
   import "./global.css";
   let { children } = $props();
 
@@ -34,9 +34,6 @@
   <Navbar />
 </div>
 {@render children()}
-<div class="footer-container">
-  <Footer />
-</div>
 
 <style>
   @import "./global.css";
@@ -77,17 +74,6 @@
     width: 100%;
     position: sticky;
     top: 0;
-    left: 0;
-    right: 0;
-    margin: 0;
-    z-index: 100;
-  }
-  .footer-container {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-    padding: 0;
-    bottom: 0;
     left: 0;
     right: 0;
     margin: 0;
